server/requirements.txt
mcp[cli]
RPi.GPIO
